Change decimal type
- I don't really need to use doubles, and using float allows me to avoid a pair of casts

+++ app/src/main/java/com/zomaotoko/randomnumbers/generators/DecimalGenerator.kt
@@ -3,9 +3,9 @@ package com.zomaotoko.randomnumbers.generators
 import java.util.concurrent.ThreadLocalRandom
 
 
-class DecimalGenerator(private val lowerBound: Double, private val upperBound: Double, private val digits: Int) : Generator {
+class DecimalGenerator(private val lowerBound: Float, private val upperBound: Float, private val digits: Int) : Generator {
 
-    override fun getRandomNumber() = ThreadLocalRandom.current().nextDouble(lowerBound, upperBound)
+    override fun getRandomNumber() = ThreadLocalRandom.current().nextDouble(lowerBound.toDouble(), upperBound.toDouble())
 
     override fun getRandomNumberString() = String.format("%.${digits}f", getRandomNumber())
 }
